(wip) 코드 분리
