+++ eslint.config.mjs
@@ -4,7 +4,7 @@ export default [
         files: ["**/*.js"],
         languageOptions: {
             ecmaVersion: "latest",
-            sourceType: "script",
+            sourceType: "module",
             globals: {
                 window: "readonly",
                 document: "readonly",
