+++ index.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
     <head>
         <meta charset="UTF-8" />
@@ -45,6 +45,6 @@
             </section>
         </main>
 
-        <script src="main.js"></script>
+        <script type="module" src="main.js"></script>
     </body>
 </html>
