+++ package.json
@@ -1,23 +1,24 @@
 {
-  "name": "rhythm-game-option-simulator",
-  "version": "1.0.0",
-  "description": "A web-based tool to simulate rhythm game options such as note speed, sudden, and hidden effects.",
-  "main": "main.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/JourneyCherry/rhythm-game-option-simulator.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/JourneyCherry/rhythm-game-option-simulator/issues"
-  },
-  "homepage": "https://github.com/JourneyCherry/rhythm-game-option-simulator#readme",
-  "devDependencies": {
-    "eslint": "^9.39.1"
-  }
+    "name": "rhythm-game-option-simulator",
+    "version": "1.0.0",
+    "description": "A web-based tool to simulate rhythm game options such as note speed, sudden, and hidden effects.",
+    "main": "main.js",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/JourneyCherry/rhythm-game-option-simulator.git"
+    },
+    "type": "module",
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "bugs": {
+        "url": "https://github.com/JourneyCherry/rhythm-game-option-simulator/issues"
+    },
+    "homepage": "https://github.com/JourneyCherry/rhythm-game-option-simulator#readme",
+    "devDependencies": {
+        "eslint": "^9.39.1"
+    }
 }
